% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Add a path to the shiny app}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Name of the library}

\item{pkgname}{Name of the package}
}
\description{
Add a path to the shiny app
}
\examples{
.onLoad(libname, "FootballeuR")
}
