% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN_Python}
\alias{KNN_Python}
\title{Title}
\usage{
KNN_Python(indiv_prediction, n = Int(11), p = Int(3))
}
\arguments{
\item{indiv_prediction}{Vecteur comprennant l’individu à prédire}

\item{n}{Permet de sélectionner le nombre de voisin}

\item{p}{Défini le p dans la norme pour la p-norme}
}
\value{
La fonction renvoie un vecteur contenant :
$prediction : Renvoie une valeur boolean sur la succes de l’expédition
$proba : Renvoie le vecteur contenant la probabilité d’appartenir à chaque classe
$temps : Renvoie le temps mis pour l'encodage et la prédiction
}
\description{
Title
}
\examples{
KNN_Python(c("Ama Dablam","Autumn","France","Climber",2025,1,1,1,1,1,1,1), n = 11 , p = 3)
}
