% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{RandomForest_Python}
\alias{RandomForest_Python}
\title{Title}
\usage{
RandomForest_Python(indiv_prediction, n = Int(2))
}
\arguments{
\item{indiv_prediction}{Vecteur comprennant l’individu à prédire}

\item{n}{Nombre d'estimateur dans l’algorithme de la Random Forest}
}
\description{
Title
}
